refactor: MainApplication의 패키지 위치를 ui에서 상위로 이동
